+++ src/core/library/macroLibraryManager.ts
@@ -8,13 +8,15 @@ export const USER_HOME_TOKEN = '${userHome}';
 export const WORKSPACE_TOKEN = '${workspaceFolder}';
 
 export class MacroLibraryManager implements vscode.Disposable {
+  private readonly configKey: string;
   public readonly libraries: Lazy<readonly MacroLibrary[]>;
   private readonly onDidChangeConfigDisposable: vscode.Disposable;
 
   constructor() {
-    this.libraries = new Lazy(() => MacroLibraryManager.getLibraries());
+    this.configKey = SOURCE_DIRS_CONFIG;
+    this.libraries = new Lazy(() => this.getLibraries());
     this.onDidChangeConfigDisposable = vscode.workspace.onDidChangeConfiguration((e) => {
-      if (e.affectsConfiguration(SOURCE_DIRS_CONFIG)) {
+      if (e.affectsConfiguration(this.configKey)) {
         this.libraries.reset();
       }
     });
@@ -22,6 +24,11 @@ export class MacroLibraryManager implements vscode.Disposable {
 
   dispose() {
     this.onDidChangeConfigDisposable.dispose();
+  }
+
+  private getConfigValue(folder?: vscode.WorkspaceFolder) {
+    return vscode.workspace.getConfiguration(undefined, folder)
+      .get<string[]>(this.configKey, []);
   }
 
   public async getFiles(): Promise<Record<string, vscode.Uri[]>> {
@@ -35,27 +42,24 @@ export class MacroLibraryManager implements vscode.Disposable {
     return allFiles;
   }
 
-  private static getLibraries(): MacroLibrary[] {
+  private getLibraries(): MacroLibrary[] {
     const uniqueRoots = new Set<string>();
+    const { workspaceFolders } = vscode.workspace;
 
-    getConfigValue().forEach((path) => expandSourceDirPath(
-      path, ...vscode.workspace.workspaceFolders ?? []));
+    for (const path of this.getConfigValue()) {
+      const expandedPaths = [];
+      if (path.includes(USER_HOME_TOKEN)) {
+        expandedPaths.push(path.replace(USER_HOME_TOKEN, os.homedir()));
+      } else if (path.includes(WORKSPACE_TOKEN)) {
+        workspaceFolders?.forEach((folder) => {
+          expandedPaths.push(path.replace(WORKSPACE_TOKEN, folder.uri.fsPath));
+        });
+      } else {
+        expandedPaths.push(path);
+      }
+      expandedPaths.forEach((p) => uniqueRoots.add(p.replace(/[/\\]+$/, '')));
+    }
 
     return [...uniqueRoots].map((root) => new MacroLibrary(vscode.Uri.file(root)));
-
-    function getConfigValue(folder?: vscode.WorkspaceFolder) {
-      return vscode.workspace.getConfiguration(undefined, folder)
-        .get<string[]>(SOURCE_DIRS_CONFIG, []);
-    }
-
-    function expandSourceDirPath(path: string, ...workspaceFolders: vscode.WorkspaceFolder[]) {
-      if (path.includes(USER_HOME_TOKEN)) {
-        uniqueRoots.add(path.replace(USER_HOME_TOKEN, os.homedir()));
-      } else if (path.includes(WORKSPACE_TOKEN)) {
-        workspaceFolders.forEach((folder) => {
-          uniqueRoots.add(path.replace(WORKSPACE_TOKEN, folder.uri.fsPath));
-        });
-      }
-    }
   }
 }