+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.3
 - Upgrade minimum VS Code version to 1.99
+- Fix: `sourceDirectories` ignores non-tokenized paths.
 
 ## 0.3.2
 - Fix: Trailing `/` affecting `vscode-remote` execution.